requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.1
mysql-connector-python==8.3.0
requests==2.31.0
boto3==1.34.33
ydb==3.8.0
soundfile==0.12.1
